Переместил стротейдж в сервис. Исправил имя исключения
