--- src/main/java/ru/yandex/practicum/filmorate/exception/search/SearchEmptyRequestParam.java
@@ -1,7 +1,0 @@
-package ru.yandex.practicum.filmorate.exception.search;
-
-public class SearchEmptyRequestParam extends RuntimeException {
-    public SearchEmptyRequestParam() {
-        super("Request param is empty");
-    }
-}
