+++ src/main/java/ru/yandex/practicum/filmorate/exception/search/SearchEmptyRequestParamException.java
@@ -1,0 +1,7 @@
+package ru.yandex.practicum.filmorate.exception.search;
+
+public class SearchEmptyRequestParamException extends RuntimeException {
+    public SearchEmptyRequestParamException() {
+        super("Request param is empty");
+    }
+}
